+++ src/views/pages/ViewPayslip.vue
@@ -123,7 +123,7 @@
                               <!-- {{ formatPrice(Math.round(item.potongan_bon)) }} -->
 
                               <v-btn
-                                v-if="item.potongan_bon == 0"
+                                v-if="item.potongan_bon == 0 && item.sisa_bon > 0"
                                 color="blue darken-1"
                                 small
                                 class="mr-3 elevation-0"
@@ -144,7 +144,20 @@
                             </td>
                             <td>Pot Lain</td>
                             <td>
-                              {{ formatPrice(Math.round(item.potongan_lain)) }}
+                              <v-btn
+                                v-if="item.potongan_lain == 0 "
+                                color="blue darken-1"
+                                small
+                                class="mr-3 elevation-0"
+                                @click="openDialogPotongan(item)"
+                                >{{
+                                  formatPrice(Math.round(item.potongan_bon))
+                                }}</v-btn
+                              >
+                              <div v-else>
+                                {{ formatPrice(Math.round(item.potongan_lain)) }}
+                              </div>
+                              
                             </td>
                           </tr>
                           <tr>
@@ -233,7 +246,7 @@
               color="blue darken-1"
               small
               class="mr-3 elevation-0"
-              @click="openDialogBon(item)"
+              @click="openDialogPotongan(item)"
               >{{ formatPrice(Math.round(item.potongan_bon)) }}</v-btn
             >
             <div v-else>{{ formatPrice(Math.round(item.potongan_bon)) }}</div>
@@ -258,6 +271,7 @@
     </v-card>
     <v-dialog v-model="dialogPay" max-width="600">
       <v-card>
+        <v-card-title><div>Pembayaran Bon Pinjaman</div></v-card-title>
         <v-card-text>
           <v-currency-field
             color="grey darken-2"
@@ -286,6 +300,32 @@
         </v-card-actions>
       </v-card>
     </v-dialog>
+    <v-dialog v-model="dialogPotonganLain" max-width="600">
+      <v-card>
+        <v-card-title>Potongan</v-card-title>
+        <v-card-text>
+          <v-currency-field
+            color="grey darken-2"
+            :decimal-length="0"
+            prefix="Rp"
+            filled
+            v-bind="currency_config"
+            v-model.trim="potongan_lain"
+            class="currency-input pa-0 ma-0 font-md"
+            label="Nominal Potongan"
+          />
+        </v-card-text>
+        <v-card-actions>
+          <div class="flex-grow-1"></div>
+          <v-btn class="elevation-0 grey darken-2" dark @click="dismisDialogPotonganLain"
+            >Batal</v-btn
+          >
+          <v-btn class="elevation-0 primary" @click.stop="savePotonganLain"
+            >Simpan</v-btn
+          >
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
   </v-container>
 </template>
   
@@ -302,6 +342,8 @@
         expanded: [],
         selected: [],
         dialogPay : false,
+        dialogPotonganLain : false,
+        potongan_lain : 0,
         loan : {},
         dataPayslip : {},
         headers: [
@@ -363,7 +405,7 @@
       };
     },
     methods: {
-      ...mapActions(['updatePayslipWithBon']),
+      ...mapActions(['updatePayslipWithBon', 'updatePayslipWithPotonganLain']),
       formatPrice(value) {
         return formatPrice(value);
       },
@@ -373,7 +415,24 @@
       formatDateUtils(val){
         return  formatDate(val, 'short-date')
       },
-  
+      savePotonganLain(){
+        const data = {
+          idPayslip : this.dataPayslip.id, 
+          potongan_lain: this.potongan_lain,
+          jenis_potongan: "produksi",
+        };
+        this.updatePayslipWithPotonganLain(data)
+        // console.log(data)
+        this.dismisDialogPotonganLain()
+      },
+      openDialogPotongan(item){
+        // console.log(item)
+        this.dataPayslip =item
+        this.dialogPotonganLain = true
+      },
+      dismisDialogPotonganLain() {
+        this.dialogPotonganLain = false;
+      },
       formatDateBulan(date) {
         date = new Date(date);
         var bulan = [